glsl-1.00: Eliminate occurences of highp in FS tests
Since we cannot be sure if highp is available in an implementation's
fragment shader, replace all occurences of highp with lowp and mediump.

(Note: High precision is not available in the GLSL 1.00 fragment shader unless
GL_FRAGMENT_PRECISION_HIGH is 1.)

+++ tests/spec/glsl-1.00/compiler/precision-qualifiers/default-precision-bool-01.frag
@@ -15,7 +15,7 @@
 
 #version 100
 
-precision highp bool;
+precision lowp bool;
 
 float f() {
 	return 0.0;
