+++ tests/spec/glsl-1.00/compiler/precision-qualifiers/precision-function-param-01.frag
@@ -11,6 +11,6 @@
 
 #version 100
 
-highp float f(lowp int x) {
+lowp float f(lowp int x) {
 	return 0.0;
 }
