+++ tests/spec/glsl-1.00/compiler/precision-qualifiers/precision-global-01.frag
@@ -11,8 +11,8 @@
 
 #version 100
 
-uniform lowp float x;
-in highp float y;
+uniform mediump float x;
+in lowp float y;
 
 float f() {
 	return 0.0;
