+++ tests/spec/glsl-1.00/compiler/precision-qualifiers/precision-literal-const-02.frag
@@ -11,6 +11,6 @@
 #version 100
 
 float f() {
-	highp float x = highp 0.0;
+	lowp float x = lowp 0.0;
 	return 0.0;
 }
