+++ tests/spec/glsl-1.00/compiler/precision-qualifiers/precision-local-01.frag
@@ -12,6 +12,6 @@
 #version 100
 
 float f() {
-	highp float x = 0.0;
+	lowp float x = 0.0;
 	return 0.0;
 }
