+++ tests/spec/glsl-1.00/compiler/precision-qualifiers/precision-struct-02.frag
@@ -12,7 +12,7 @@
 
 #version 100
 
-highp struct s {
+lowp struct s {
 	float a;
 };
